comments/requirements.txt
blinker==1.7.0
click==8.1.7
flask==2.2
gunicorn==21.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
packaging==23.2
protobuf==4.21.12
werkzeug==2.2.2
zipp==3.17.0
flask-login==0.6.3
flask-session==0.5.0
flask-sqlalchemy==3.0.2
pymysql==1.1.0